quorum_public

Module cli

Main entry point for the server cli Commands This is where commands are defined properly, and where command functions are called.

Modules

help

Main cli file used for getting help with server commands

server

test

user

Structs

Command 🔒

Constants

SESSION_TIMEOUT_MINS 🔒

Functions

dispatch 🔒

Dispatches the parsed command to the appropriate handler function.

parse 🔒

Parses a command string into a Command struct.

require_admin

Checks if a command required admin privileges.

spawn_cli

Spawns the CLI in a separate blocking thread.

unknown 🔒

Prints an error message for unknown commands.