quorum_public::cli

Module server

Functions

login

Handles user login

make_admin

Handles making a user an admin

prompt 🔒

Prompts the user for input with a given label and returns the trimmed input as a String.

prompt_password 🔒

Prompts the user for a password with a given label and returns the input as a String.

signup

Handles user signup