quorum_public::db::queries

Module auth

Authentication related database queries

This file contains functions for Creating an account (signup), login, token management and deleteing a user account.

Functions

delete_user_by_id

Deletes a user from the database by their ID

make_admin

Promotes a user to admin status in the database

revoke_refresh_token

Revokes a refresh token, preventing it from being used for future authentication

signup_user

Creates a new user account in the database

store_refresh_token

Stores a refresh token in the database for a specific user

update_user_profile

Updates a user’s profile information in the database

validate_refresh_token

verify_user_credentials

Get’s user’s data to allow them to be verified